server/requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
psycopg2-binary==2.9.4
pytest-django==4.5.2
pytest==7.2.0
Pillow==9.0.0
pytest-cov==4.0.0
flake8==5.0.4
djangorestframework-simplejwt==5.0.0
requests==2.28.1
django-cors-headers==3.13.0